lib: use /~/host endpoint if cookie not legible
In the future, Eyre will serve its cookies with the HttpOnly flag to
protect against theft. urbit/urbit#6561 adds a /~/host endpoint we can
use to discover the name of the ship we're talking to. So, if the cookie
is illegible to us, make a request to that endpoint instead.

Changes herein mostly copied from urbit/urbit#5918.

Co-authored-by: pkova <pyry.kovanen@gmail.com>