lib: use nockjs from npm instead of local files
Since the nockjs api has changed lightly since the code herein was
written, we have to make some small adjustments.

Note that this also changes the interface for the .err() subscription
interface callback, to bring it in line with the changed to the .event()
interface we introduced during earlier cleanup.