lib: update for eyre jam-channels api changes
We no longer use a special path (/~/channel-jam), but instead make use
of request headers to indicate what mode to open a channel in, and how
to interpret a channel event.

This required loosening up the headers type a little bit, making changes
to the fetchOptions, making them dynamic rather than static.