lib: resolve poke() calls eagerly, never throw
Previously, the poke() function was asynchronous on _both_ its PUT
request _and_ the poke-n/ack. In the nack case, it would *reject* the
latter promise instead of resolving it, causing the whole to throw an
error. Due to implementation mishap, this would always throw undefined.

Here, we update this behavior, so that it's asynchronous on _only_ the
PUT request, and returns the event id as soon as that resolves. The
callbacks passed as arguments still get called as normal.

This misbehavior was discovered while finishing the implementation of
the "handle poke nack" test, whose final form has been included.

Note that this is a subtle breaking change. Clients who were awaiting on
their poke() call will now see different/faster behavior, and no longer
need to fold it into a catch block.