lib: start restructuring connection logic
There were shenanigans aronud .eventSource getting recursively called
into by .sendNounToChannel. This whole structure was unnecessarily
recursive and hard to follow.

Here we flatten it out a little bit, making .eventSource responsible for
continuing after an initialization poke, in the "not yet initialized"
case.

This means that developers using this library need to explicitly call
.eventSource() in order to establish a connection. This probably makes
for a better API, although we should add a success/failure return value
to it in the near future.