lib: make sendNounsToChannel() match eyre's api
Eyre takes a list of commands for each incoming PUT. Here, we update
sendNounToChannel() into sendNounsToChannel(), taking a variable number
of arguments, which get converted into "real" Nouns inside of the
function, using nockjs's dwim logic. (Of course, this means callers are
still free to pass pre-built Nouns also.) This way, we don't end up with
Noun construction boilerplate all over the place, and this lets us send
the list of empty commands, which we want to use for channel
initialization.