lib: scry out nouns, not json, by default
Necessarily includes logic for turning the jambuffer in the response,
into a proper Noun object.

Only somewhat surprisingly, we need to reverse the byte order in the
response, even though to accurately mock it, we don't need to reverse
the bytes from the jam generated by nockjs.

It would be most prudent to manually test this in the browser, to make
sure we didn't trip ourselves up over endianness shenanigans here.