api: cleaned up setup and exports
This makes sure we re-export important functions from @urbit/nockjs. We
also removed the URL prefixing which was causing issues with passing empty
string as the URL. Finally, we reconfigured the setup so that the Urbit
instance returned is immediately available and all interactions wait for
the connection to be established.